Revert MsgBox to run on the UI thread. Fix CallbackCreate callback calling when current thread is critical.

+++ Keysharp.Core/Builtins/Buffer.cs
@@ -47,7 +47,7 @@ namespace Keysharp.Builtins
 
 				if (val > size)
 				{
-					var newptr = Marshal.AllocHGlobal((int)val);
+					var newptr = Marshal.AllocHGlobal((nint)val);
 
 					if (_ptr != null)
 					{
